+++ topics.js
@@ -778,7 +778,6 @@ async function handleTopicAvatars(topics) {
 
         const avatarUrl = await getAuthorAvatarUrl(author, authorProfileUrl);
         if (avatarUrl?.length) {
-            authorAvatar.alt = author;
             authorAvatar.src = avatarUrl;
             authorAvatar.onerror = `this.onerror=null; this.src='${defaultAvatarUrl}';`;
         }